push des évaluation et autres

+++ CheckMyStar.Bll.Models/StarCriteriaDetailModel.cs
@@ -4,7 +4,7 @@
     {
         public int CriterionId { get; set; }
         public string? Description { get; set; }
-        public decimal BasePoints { get; set; }
+        public int BasePoints { get; set; }
         public string? TypeCode { get; set; }
         public string? TypeLabel { get; set; }
     }
